+++ src/SqlStreamStore.HAL.Tests/StreamMetadataTests.cs
@@ -1,7 +1,5 @@
 ﻿namespace SqlStreamStore.HAL.Tests
 {
-    using System;
-    using System.Collections;
     using System.Collections.Generic;
     using System.Linq;
     using System.Net;
