+++ src/SqlStreamStore.HAL/ReadAllStreamMiddleware.cs
@@ -1,10 +1,9 @@
 namespace SqlStreamStore.HAL
 {
-    using System.Net.Http;
-    using System.Threading.Tasks;
     using Microsoft.Owin;
     using Microsoft.Owin.Builder;
     using Owin;
+    using SqlStreamStore.HAL.Resources;
     using MidFunc = System.Func<System.Func<System.Collections.Generic.IDictionary<string, object>,
             System.Threading.Tasks.Task
         >, System.Func<System.Collections.Generic.IDictionary<string, object>,
@@ -16,12 +15,11 @@ namespace SqlStreamStore.HAL
         public static MidFunc UseStreamStore(IReadonlyStreamStore streamStore)
         {
             var allStream = new AllStreamResource(streamStore);
+            var allStreamMessages = new AllStreamMessageResource(streamStore);
 
             var builder = new AppBuilder()
                 .MapWhen(IsAllStream, inner => inner.Use(GetStream(allStream)))
-                .MapWhen(IsStreamOptions, inner => inner.Use(GetStreamOptions))
-                .MapWhen(IsAllStreamMessage, inner => inner.Use(GetStreamMessage(allStream)))
-                .MapWhen(IsStreamMessageOptions, inner => inner.Use(GetStreamMessageOptions));
+                .MapWhen(IsAllStreamMessage, inner => inner.Use(GetStreamMessage(allStreamMessages)));
 
             return next =>
             {
@@ -34,14 +32,9 @@ namespace SqlStreamStore.HAL
         private static bool IsAllStream(IOwinContext context)
             => context.IsGetOrHead() && context.Request.Path.IsAllStream();
 
-        private static bool IsStreamOptions(IOwinContext context)
-            => context.IsOptions() && context.Request.Path.IsAllStream();
-
         private static bool IsAllStreamMessage(IOwinContext context)
             => context.IsGetOrHead() && context.Request.Path.IsAllStreamMessage();
 
-        private static bool IsStreamMessageOptions(IOwinContext context)
-            => context.IsOptions() && context.Request.Path.IsAllStreamMessage();
 
         private static MidFunc GetStream(AllStreamResource allStream) => next => async env =>
         {
@@ -57,23 +50,11 @@ namespace SqlStreamStore.HAL
             }
         };
         
-        private static MidFunc GetStreamOptions => next => env =>
-        {
-            var context = new OwinContext(env);
-            
-            context.SetStandardCorsHeaders(
-                HttpMethod.Get, 
-                HttpMethod.Head, 
-                HttpMethod.Options);
-
-            return Task.CompletedTask;
-        };
-
-        private static MidFunc GetStreamMessage(AllStreamResource allStream) => next => async env =>
+        private static MidFunc GetStreamMessage(AllStreamMessageResource allStreamMessages) => next => async env =>
         {
             var context = new OwinContext(env);
 
-            var response = await allStream.GetMessage(
+            var response = await allStreamMessages.GetMessage(
                 new ReadAllStreamMessageOptions(context.Request),
                 context.Request.CallCancelled);
 
@@ -82,18 +63,5 @@ namespace SqlStreamStore.HAL
                 await context.WriteHalResponse(response);
             }
         };
-        
-        private static MidFunc GetStreamMessageOptions => next => env =>
-        {
-            var context = new OwinContext(env);
-            
-            context.SetStandardCorsHeaders(
-                HttpMethod.Get, 
-                HttpMethod.Head, 
-                HttpMethod.Options);
-
-            return Task.CompletedTask;
-        };
-
     }
 }
