+++ src/SqlStreamStore.HAL/ReadStreamMiddleware.cs
@@ -5,6 +5,7 @@ namespace SqlStreamStore.HAL
     using Microsoft.Owin;
     using Microsoft.Owin.Builder;
     using Owin;
+    using SqlStreamStore.HAL.Resources;
     using MidFunc = System.Func<System.Func<System.Collections.Generic.IDictionary<string, object>,
             System.Threading.Tasks.Task
         >, System.Func<System.Collections.Generic.IDictionary<string, object>,
@@ -16,12 +17,11 @@ namespace SqlStreamStore.HAL
         public static MidFunc UseStreamStore(IStreamStore streamStore)
         {
             var streams = new StreamResource(streamStore);
-            
+            var streamMessages = new StreamMessageResource(streamStore);
+
             var builder = new AppBuilder()
-                .MapWhen(IsStreamMessage, inner => inner.Use(GetStreamMessage(streams)))
-                .MapWhen(IsStreamMessageOptions, inner => inner.Use(GetStreamMessageOptions))
-                .MapWhen(IsStream, inner => inner.Use(GetStream(streams)))
-                .MapWhen(IsStreamOptions, inner => inner.Use(GetStreamOptions));
+                .MapWhen(IsStreamMessage, inner => inner.Use(GetStreamMessage(streamMessages)))
+                .MapWhen(IsStream, inner => inner.Use(GetStream(streams)));
 
             return next =>
             {
@@ -34,22 +34,16 @@ namespace SqlStreamStore.HAL
         private static bool IsStream(IOwinContext context)
             => context.IsGetOrHead() && context.Request.Path.IsStream();
 
-        private static bool IsStreamOptions(IOwinContext context)
-            => context.IsOptions() && context.Request.Path.IsStream();
-
         private static bool IsStreamMessage(IOwinContext context)
             => context.IsGetOrHead() && context.Request.Path.IsStreamMessage();
 
-        private static bool IsStreamMessageOptions(IOwinContext context)
-            => context.IsOptions() && context.Request.Path.IsStreamMessage();
-
-        private static MidFunc GetStream(StreamResource stream) => next => async env =>
+        private static MidFunc GetStream(StreamResource streams) => next => async env =>
         {
             var context = new OwinContext(env);
 
             var options = new ReadStreamOptions(context.Request);
 
-            var response = await stream.GetPage(options, context.Request.CallCancelled);
+            var response = await streams.GetPage(options, context.Request.CallCancelled);
 
             using(new OptionalHeadRequestWrapper(context))
             {
@@ -57,27 +51,13 @@ namespace SqlStreamStore.HAL
             }
         };
 
-        private static MidFunc GetStreamOptions => next => env =>
-        {
-            var context = new OwinContext(env);
-
-            context.SetStandardCorsHeaders(
-                HttpMethod.Get,
-                HttpMethod.Head,
-                HttpMethod.Options,
-                HttpMethod.Post,
-                HttpMethod.Delete);
-
-            return Task.CompletedTask;
-        };
-
-        private static MidFunc GetStreamMessage(StreamResource stream) => next => async env =>
+        private static MidFunc GetStreamMessage(StreamMessageResource streamMessages) => next => async env =>
         {
             var context = new OwinContext(env);
 
             var options = new ReadStreamMessageOptions(context.Request);
 
-            var response = await stream.GetMessage(options, context.Request.CallCancelled);
+            var response = await streamMessages.GetMessage(options, context.Request.CallCancelled);
 
             using(new OptionalHeadRequestWrapper(context))
             {
