+++ src/SqlStreamStore.HAL/Resources/DeleteStreamMetadataOptions.cs
@@ -1,4 +1,4 @@
-namespace SqlStreamStore.HAL
+namespace SqlStreamStore.HAL.Resources
 {
     using System;
     using System.Threading;
