+++ src/SqlStreamStore.HAL/Resources/IResource.cs
@@ -1,0 +1,9 @@
+﻿namespace SqlStreamStore.HAL.Resources
+{
+    using System.Net.Http;
+
+    internal interface IResource
+    {
+        HttpMethod[] Options { get; }
+    }
+}
