+++ src/SqlStreamStore.HAL/Resources/ReadAllStreamMessageOptions.cs
@@ -1,4 +1,4 @@
-namespace SqlStreamStore.HAL
+namespace SqlStreamStore.HAL.Resources
 {
     using System;
     using System.Linq;
